+++ include/xsk/arc/compiler.hpp
@@ -21,8 +21,8 @@ class compiler
     std::unordered_map<std::string, expr const*> constants_;
     std::string animtree_;
     sourcepos debug_pos_;
-    u32 index_;
-    u32 label_idx_;
+    usize index_;
+    usize label_idx_;
     bool can_break_;
     bool can_continue_;
     bool developer_thread_;
